lint: move schema to directory, make graph part of schema
