+++ src/schema/graph/input/Pairs.ts
