+++ src/schema/graph/output/Pairs.ts
